refactor: Move edit tab hooks to appropriate level in the file structure
