+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/EditTab/OptionListEditor/hooks/index.ts
@@ -1,3 +1,1 @@
-export { useTextResourcesForLanguage } from './useTextResourcesForLanguage';
-export { useHandleUpdateTextResource } from './useHandleUpdateTextResource';
 export { useConcatOptionListLabels } from './useConcatOptionListLabels';
