+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/EditTab/hooks/useHandleUpdateTextResource.test.ts
